+++ src/components/Form/ExpirationDateForm.tsx
@@ -1,7 +1,8 @@
-import { useState, useEffect } from "react";
+import { useState } from "react";
 import Input from "./Input";
 import FormElement from "../common/FormElement";
 import { CardFormProps } from "./CardNumberForm";
+import useExpirationDateForm from "../../hooks/useExpirationDateForm";
 
 interface ExpirationDateFormProps extends CardFormProps {
   expirationDate: string[];
@@ -18,47 +19,14 @@ const ExpirationDateForm = ({
   onValidation,
   onFocus,
 }: ExpirationDateFormProps) => {
-  const [, setAllInputValid] = useState(false);
-  const [errorMessage, setErrorMessage] = useState("");
-  const [inputValidity, setInputValidity] = useState({
-    0: false,
-    1: false,
-  });
-
   const [isFocused, setIsFocused] = useState(false);
+  const { errorMessage, setErrorMessage, validateMonth, validateYear, updateInputValidity } =
+    useExpirationDateForm(onValidation);
 
   const handleFocus = (type: string) => {
     onFocus(type);
     setIsFocused(true);
   };
-
-  const validateMonth = (value: string) => {
-    const regex = /^(0[1-9]|1[0-2])$/;
-
-    return regex.test(value);
-  };
-
-  const validateYear = (value: string) => {
-    const intValue = Number(value);
-    return !isNaN(intValue) && intValue >= 24 && intValue <= 29;
-  };
-
-  // NOTE: 각 입력 필드의 유효성 검사 결과를 업데이트
-  const updateInputValidity = (index: number, isValid: boolean) => {
-    setInputValidity((prevValidities) => ({
-      ...prevValidities,
-      [index]: isValid,
-    }));
-  };
-
-  useEffect(() => {
-    const allValid = Object.values(inputValidity).every((isValid) => isValid);
-
-    setAllInputValid(allValid);
-    setErrorMessage(allValid ? "" : "만료 기한을 올바르게 입력해주세요.");
-
-    if (onValidation) onValidation(allValid);
-  }, [inputValidity]);
 
   const inputs = Array.from({ length: inputCount }, (_, index) => (
     <Input
